requirements.txt
PyGithub==1.25.0